+++ Changelog.md
@@ -4,6 +4,8 @@
 
 - mkvtoolnix 48
 - ec3 file extension support for eac3 files.
+- R210/V210 video output using ffmpeg.
+
 
 2.1.3.7 Beta
 ============